Initial commit: Add OurWorldInData.org data source meta-data
This is intended to become a half-automated data source processing
mechanism to generate your own plots from public data available
about the covid-19 (Corona Virus Disease started 2019) crisis
as the situation unfolds.

That is, the aim is to be able to half-automatically download
new data and replot it to image files; but it should also *not*
download/replot anything when a data source can be tested to
not have been updated since last time checked, and keep every
version of source data (and possibly graphs?) ever encountered.
(e.g., to be able to diff them against each other to spot
unexpected changes or something.)

The first few commits will be back-dated as I go on importing
existing data (from file-copy-with-date-in-filename versions)
into git.

<https://ourworldindata.org/coronavirus> was the first data source
for which I've written a converter (in Perl) for easy use with gnuplot.