JHU: converter v2.3: Fix timestamp to 1s before next day UTC
gnuplot itself does not seem to make a difference (as far as date
stringification on the x-axis is concerned). But it already caused
confusion on my part when looking at a processed data file to find out
how current it was, and the time stamp of, e.g., 2020-03-18 00:00:00 UTC
mislead me to beleave this was rather -17's data, although it was the
full -18's data in fact. So generate 2020-03-18 23:59:59 UTC, instead.