OWID: Extract download.sh from Makefile; invoke generator
With this, we can have the download for full_data.csv
in a make target with a different name (download-full_data.csv),
which should avoid running the recipe for full_data_for_gnuplot.csv
on *every* `make update` run.

(When renaming the target from the resulting file to a separate
phony target, all the uses of $@ would have needed to be changed
to something different. And per-target variable assignment could
also not be used as it would apply to prerequisites, too, so would
not be a semantically correct solution.)

Besides, it's possible but quite unreadable (unmaintainable?)
to embed large portions of shell script into a Makefile, so
the solution with having a separate shell script should be better,
now.