OWID: examples: Start converting to common include file
This should both make the actual plot files shorter,
and provide additional functionality not present before --
like PNG file output. For the future, there are already hooks
to put additional stuff in-between setup and plot commands.

Use that like so:

    covid-19-canvon/OurWorldInData.org$ gnuplot -c \
        examples/1.1-totalCases_Welt_Deutschland.gnuplot.txt \
	png Taeteretae.png

This should make it possible to script plot generation, soon.