OWID: Fix URL to what will still be updated
It seems the URL changed, from data/full_data.csv
to data/ecdc/full_data.csv, possibly with a change from
getting the numbers from the WHO to getting it from "ECDC".

Empty fields seem to be filled up with 0s, now, and past numbers
differ by, e.g., 1 person here and there... But the general structure
seems to be the same, so everything should continue to work, now with
current numbers, again.