JHU: examples: Add logscale versions 2.{2,3}b (of 2.{2,3})
This is as the total deaths gets more interesting as it was
recently reported that Spain's value would be surpassing that of China,
after Italy already did so some time ago. So maybe we can have a look
at how/wheather things differ between logscale plots in a different way
as between the linear scale plots already available.

(On the other hand, it makes more possibly boring graphs through which
to step through on cursory correctness verification... Plus, the order
2.1-*, 2.1b-*, 2.2-*, 2.2b-*, ... works in the `LC_COLLATE=C.UTF-8`
locale, but doesn't in, e.g., `en_GB.UTF-8`, as it seems to skip
punctuation and, then, compare alphabetically, instead... So the order
will be wrong when simply doing a shell glob.)