JHU: Makefile: Add targets examples/%.png, plot-examples-inplace
.., for easier testing.

That is, now a single PNG can be built from current data
to always-same-place (alongside the gnuplot source),
which should make it easier to build/view multiple times
in a row. Especially, no archive/plots/YYYYMMDD_HHMMSS/
directories will have to be removed manually after testing.

For ease of testing everything, the `make plot-examples-inplace`
is provided.

* .gitignore: Ignore resulting examples/*.png