gnuplot: Stop setting terminal via call argument pass-through
Before, `gnuplot -c foo.gnuplot.txt png outfile.png` was used.
After, `gnuplot -e "call '...' 'png' 'outfile.png'" foo.gnuplot.txt`
is used.

This might seem less good at first sight, because of the string
interpolation needed. But it frees us from having to "pass through"
the first 3 `call` arguments from each and every script
to the `common.gnuplot.inc.txt`. Furthermore, it should now be possible
to `load` or `call` an example (gnuplot instruction file)
without explicitly specifying a terminal, without getting the terminal
reset to qt (which will not work on, e.g., gnuplot on Termux
on Android and can not easily be fixed by calling again due to
the custom header guard / "include once and only once" behaviour,
in this repository.)

* Add terminal.gnuplot: Pull (some of) the `set terminal ...` code
  out of `common.gnuplot.inc.txt`. Use `exit error "foo"` for the
  first time, when argument checking fails.
  It just sets the given terminal with a size of 1024x768, and,
  optionally, sets the output file.
* common.gnuplot.inc.txt: See above. Also, drop the @ARG3 at the end,
  it wasn't used, anyway, and maybe makes things more complicated.
* */Makefile, plot_examples.sh: Use the new-style gnuplot invocation.
* */examples/*gnuplot*: Stop doing general `call` argument pass-through.
  (Only pass through to gnuplot instruction files that actually use
  the arguments themselves.)

Note that this is the first time in this repos that the .gnuplot
file name extension is used. I didn't have automatic activation
of the VIM syntax highlighting for gnuplot before, and have it now
only via a custom `~/.vim/ftdetect/gnuplot.vim`, containing:

    au BufNewFile,BufRead *.gnuplot setf gnuplot

But .gnuplot is used here (instead of, e.g., .gpi which VIM supports
out-of-the-box) to clearly communicate for which program these files
are meant, at first sight.

The .gnuplot.txt should have made it easier for graphical users
to launch a text editor on them (or have put a web server hosting them
to serve them as text so a casual visitor can have a look easily),
but there are extensionless files (Makefile), already, which will need
a manual editor invocation anyway; and it should be possible
for graphical users to associate the .gnuplot extension
with a text editor (possibly as alternative to gnuplot itself),
and maybe they'll already have gnuplot installed and an association
of .gnuplot files with gnuplot. ... Thinking about this too much,
already...