gnuplot: common: Have one-week between xtics, and 7 days of mxtics
The autoscaling (autofreq?) is not that good, in that it'll choose
two-weeks between xtics, with 2 weeks of mxtics (one per week),
nowadays. We won't go lower than 1 day per mxtic due to our data points
being of 1 day granularity, and will not switch to a very much larger
time span soon, either. So set it explicitly, for now.

(See the commit diff for further related problems/thoughts...)