slide: Have count/size on the list below carousel, too
This should make it easier to correllate between carousel and list
via slide number, too. (Still `ul` is used instead of `ol`,
as having the slide number in front would collide with the
hierarchical number (e.g., 2.3) at the beginning of the plot
base name.)

To avoid confusion or to guide the eye, the get-at-raw-file
is now put into square brackets instead of parentheses,
a bit like a button.