slide: Have new history entry *before* jumping to slide
Otherwise, back button will take us to near the slide/carousel only,
and we haven't won anything (in terms of useability).

With this combined with the upcoming addition of downward links
from the plot/slide, it should be possible to jump up/down
multiple times in a row, and then use back button to jump in reverse.