slide: Bugfix: Fix wrong slideNum via click on link to slide
It turns out the use of `this.getAttribute('data-slide-to')`
didn't do what I meant. Instead of doing an auto-conversion
from text to number and getting slide 0 -> `#slideNum=1`,
I got `#slideNum=01` instead...

So, instead of trying to be general, just embed the slide index
in the onclick JavaScript, too... This should fix it.