slide: Switch from pure CSS :focus-within to JS .focus-img
That is, simulate the CSS pseudo-class :focus-within by using
JavaScript and jQuery's addClass()/removeClass(). This will
trigger the pseudo-fullscreen for the 1:1 plot image only
when the img itself receives focus, not when anchors in the slide do.

This should fix being unable to click on / follow links in a slide
from the normal page. Using those same links from the pseudo-fullscreen
does not work now, though, probably as the focus gives the class which
lets pseudo-fullscreen go away before a mouse up event on the anchor,
so there is no click event and no link activation... (?)