slide: Bugfix: Allow use of links in a slide from pseudo-fullscreen
From the commit diff / HTML comment below element attribute JS:

> Use timeout to give the links in the slide a chance
> to trigger, before removing the class will snatch them
> away from below the pointer...

Amendment:

This does not seem to be always stable, though. :(
Sometimes, especially on generated slide.html with fewer plots (?),
clicking a link in the slide in pseudo-fullscreen will alter
the URL hash, but won't always correctly jump to the target.

Simply jumping there explicitly via JS after the class has been removed
does not seem to be an option, as normally, upon exiting from
pseudo-fullscreen we should return to the same scrolling position
of the page as before entering it!

Should we probably save the hash on entering pseudo-fullscreen,
compare against hash on leaving, and, only then, jump via JS..?