slide: Try to avoid stealing all imgs' focus on slide
Achieve the desired effect, of not accidentally jump back
to pseudo-fullscreen mode via doing a slide, directly,
by removing the associated class, instead of indirectly
by going via the img's blur event handler.

This didn't help with being able to do multiple times left/right
in a row from an img that had focus, but is probably a step
in that direction.