JHU: Add convert_population.pl (Perl) & make target
..for `population_for_gnuplot.csv`.

With this, the plots could be made "per 100'000" (relative to
population), instead of "racing" each country against each other
via simple total cases. This is not implemented in the examples,
yet, though.

First time we're using a "real" CSV parser here, Text::CSV for Perl.
A bit annoying to set up in the script, but seems to work solidly!