JHU: example 1.0: For newConfirmed, use fsteps (instead of lines)
This doesn't look as good as before, but should be more correct.
(The newConfirmed are per day, and intermediary values between
each sample, as suggested by `with lines`, don't exist!)

Use fsteps instead of steps, as the time used for each data point
is the time of reporting; this should give a number for the **previous**
day.