JHU: examples 2.*: Use values-in-key
With this, the final plotted value, as absolute value and
difference-to-previous-day, will be appended to the curve names.

This looks very bad at first sight. :( But, finally, the numbers
are there without need for looking at the curve and the scales...
This should provide easy comparability with, e.g., TV news shows.